requirements.txt
decorator==4.4.2
dogpile.cache==0.9.2
redis==3.5.3
